pyproject.toml
[project]
name = "nov-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apache-airflow>=2.8.0",
    "dash>=3.3.0",
    "faker>=20.0.0",
    "jinja2>=3.1.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyspark>=3.5.0",
    "pytest>=7.0.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "typing-extensions>=4.0.0",
]
